created SAL file structure for service layer, renamed data layer, still experiencing errors connecting to mongo database
