+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/DAL/CategoryDAL/CategoryRepository.java
@@ -1,13 +1,15 @@
-package com.example.Reimbursement.Request.App.Implementation.Session;
+package com.example.Reimbursement.Request.App.DAL.CategoryDAL;
 
-import com.example.Reimbursement.Request.App.Entities.Session;
+import com.example.Reimbursement.Request.App.Entities.Category;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.SimpleMongoClientDatabaseFactory;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
-public interface SessionRepository extends MongoRepository<Session, String> {
+@Repository
+public interface CategoryRepository extends MongoRepository<Category, String> {
     MongoClient mongoClient = MongoClients.create("mongodb://localhost:27017/reimbursement-request-app");
     MongoTemplate mongoTemplate = new MongoTemplate(new SimpleMongoClientDatabaseFactory(mongoClient,
             "reimbursement-request-app"));
