+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/DAL/EmployeeDAL/EmployeeInterface.java
@@ -1,4 +1,4 @@
-package com.example.Reimbursement.Request.App.Implementation.Employee;
+package com.example.Reimbursement.Request.App.DAL.EmployeeDAL;
 
 import com.example.Reimbursement.Request.App.Entities.Employee;
