+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/DAL/EmployeeDAL/EmployeeRepository.java
@@ -1,15 +1,13 @@
-package com.example.Reimbursement.Request.App.Implementation.Category;
+package com.example.Reimbursement.Request.App.DAL.EmployeeDAL;
 
-import com.example.Reimbursement.Request.App.Entities.Category;
+import com.example.Reimbursement.Request.App.Entities.Employee;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.SimpleMongoClientDatabaseFactory;
 import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
-public interface CategoryRepository extends MongoRepository<Category, String> {
+public interface EmployeeRepository extends MongoRepository<Employee, String> {
     MongoClient mongoClient = MongoClients.create("mongodb://localhost:27017/reimbursement-request-app");
     MongoTemplate mongoTemplate = new MongoTemplate(new SimpleMongoClientDatabaseFactory(mongoClient,
             "reimbursement-request-app"));
