+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/DAL/RequestDAL/RequestInterface.java
@@ -1,4 +1,4 @@
-package com.example.Reimbursement.Request.App.Implementation.Request;
+package com.example.Reimbursement.Request.App.DAL.RequestDAL;
 
 import com.example.Reimbursement.Request.App.Entities.Request;
