+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/DAL/SessionDAL/SessionInterface.java
@@ -1,4 +1,4 @@
-package com.example.Reimbursement.Request.App.Implementation.Session;
+package com.example.Reimbursement.Request.App.DAL.SessionDAL;
 
 import com.example.Reimbursement.Request.App.Entities.Session;
