+++ Reimbursement-Request-App/src/main/resources/application.properties
@@ -3,4 +3,4 @@ spring.data.mongodb.username=rootuser
 spring.data.mongodb.password=rootpass
 spring.data.mongodb.database=reimbursement-request-app
 spring.data.mongodb.port=27017
-spring.data.mongodb.host=localhost
+spring.data.mongodb.host=b7b8ddf76faf
